+++ src/main.rs
@@ -4,8 +4,9 @@ mod custom_event;
 mod event;
 mod logger;
 mod settings;
+#[cfg(feature = "time_frame")]
+mod time_frame;
 mod ui;
-mod utils;
 
 use clap::{AppSettings, Clap};
 
@@ -52,7 +53,7 @@ struct Opts {
 fn main() {
     let opts: Opts = Opts::parse();
     #[cfg(feature = "time_frame")]
-    let mut time_frame_stat = utils::TimeStat::default();
+    let mut time_frame_stat = time_frame::TimeStat::default();
     init_logger(opts.log_level);
 
     let (mut context, mut game, mut debugger, mut event_pump) = init_gbmu(&opts);
